SAK-2324 Delete attachments when they are removed iff they are in the attachments area.
git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@3155 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ legacy/tools/src/java/org/sakaiproject/tool/content/ResourcesAction.java
@@ -3044,12 +3044,16 @@ public class ResourcesAction
 		if(found && item != null)
 		{
 			attached.remove(item);
+			
 			try
 			{
-				ContentResourceEdit edit = ContentHostingService.editResource(item.getId());
-				ContentHostingService.removeResource(edit);
-				ContentCollectionEdit coll = ContentHostingService.editCollection(item.getCollectionId());
-				ContentHostingService.removeCollection(coll);
+				if(ContentHostingService.isAttachmentResource(item.getId()))
+				{
+					ContentResourceEdit edit = ContentHostingService.editResource(item.getId());
+					ContentHostingService.removeResource(edit);
+					ContentCollectionEdit coll = ContentHostingService.editCollection(item.getCollectionId());
+					ContentHostingService.removeCollection(coll);
+				}
 			}
 			catch(Exception ignore)
 			{
